+++ sdk/ai/Azure.AI.Projects.Agents/tests/AgentsTests.cs
@@ -7,7 +7,6 @@ using System.IO;
 using System.IO.Compression;
 using System.Linq;
 using System.Threading.Tasks;
-using Castle.Core.Resource;
 using Microsoft.ClientModel.TestFramework;
 using NUnit.Framework;
 using NUnit.Framework.Internal;
@@ -42,42 +41,85 @@ public class AgentsTests : AgentsTestBase
     }
 
     [RecordedTest]
-    public async Task TestAgentCRUD()
+    [TestCase(true)]
+    [TestCase(false)]
+    public async Task TestAgentCRUD(bool useExternalAgent)
     {
         AgentAdministrationClient agentsClient = GetTestClient();
-        ProjectsAgentDefinition emptyAgentDefinition = new DeclarativeAgentDefinition(TestEnvironment.FOUNDRY_MODEL_NAME);
+        ProjectsAgentDefinition emptyAgentDefinition = useExternalAgent ? new ExternalAgentDefinition() { OtelAgentId = "foo"} :  new DeclarativeAgentDefinition(TestEnvironment.FOUNDRY_MODEL_NAME);
 
-        const string emptyPromptAgentName = "TestNoVersionAgentFromDotnetTests";
-        try
-        {
-            await agentsClient.DeleteAgentAsync(emptyPromptAgentName);
-        }
-        catch (ClientResultException)
-        {
-            // We do not have the agent to begin with.
-        }
         ProjectsAgentVersion newAgentVersion = await agentsClient.CreateAgentVersionAsync(
-            emptyPromptAgentName,
+            AGENT_NAME2,
             new ProjectsAgentVersionCreationOptions(emptyAgentDefinition)
             {
                 Metadata = { ["delete_me"] = "please " },
             });
         Assert.That(newAgentVersion?.Id, Is.Not.Null.And.Not.Empty);
 
-        ProjectsAgentRecord retrievedAgent = await agentsClient.GetAgentAsync(emptyPromptAgentName);
+        ProjectsAgentRecord retrievedAgent = await agentsClient.GetAgentAsync(AGENT_NAME2);
         Assert.That(retrievedAgent?.Id, Is.EqualTo(newAgentVersion.Name));
-
-        await agentsClient.DeleteAgentAsync(newAgentVersion.Name);
 
         ProjectsAgentVersion agentVersion = await agentsClient.CreateAgentVersionAsync(AGENT_NAME, new ProjectsAgentVersionCreationOptions(emptyAgentDefinition));
         Assert.That(AGENT_NAME, Is.EqualTo(agentVersion.Name));
-        ProjectsAgentVersion agentVersionObject_ = await agentsClient.GetAgentVersionAsync(agentName: agentVersion.Name, agentVersion: agentVersion.Version);
-        Assert.That(AGENT_NAME, Is.EqualTo(agentVersionObject_.Name));
-        Assert.That(agentVersion.Version, Is.EqualTo(agentVersionObject_.Version));
         Assert.That(agentVersion.Description, Is.Empty);
         Assert.That(agentVersion.Metadata, Is.Empty);
-        await agentsClient.DeleteAgentVersionAsync(agentName: agentVersion.Name, agentVersion: agentVersion.Version);
-        Assert.ThrowsAsync<ClientResultException>(async () => await agentsClient.GetAgentVersionAsync(agentVersion.Name, agentVersion.Version));
+        agentVersion = await agentsClient.CreateAgentVersionAsync(AGENT_NAME, new ProjectsAgentVersionCreationOptions(emptyAgentDefinition)
+        {
+            Metadata = { { "foo", "bar" } }
+        });
+        // Get Version
+        ProjectsAgentVersion agentVersionObject_ = await agentsClient.GetAgentVersionAsync(agentName: agentVersion.Name, agentVersion: agentVersion.Version);
+        ValidateDefinition(agentVersionObject_, useExternalAgent);
+        Assert.That(agentVersionObject_.Name, Is.EqualTo(AGENT_NAME));
+        Assert.That(agentVersionObject_.Version, Is.EqualTo(agentVersion.Version));
+        // Get
+        ProjectsAgentRecord agentObject_ = await agentsClient.GetAgentAsync(agentName: agentVersion.Name);
+        ValidateDefinition(agentObject_.Versions.Latest, useExternalAgent);
+        Assert.That(agentObject_.Name, Is.EqualTo(AGENT_NAME));
+        // List Agents
+        ProjectsAgentKind goodKind = useExternalAgent ? ProjectsAgentKind.External : ProjectsAgentKind.Prompt;
+        ProjectsAgentKind badKind = ProjectsAgentKind.Workflow;
+        List<ProjectsAgentRecord> records = await agentsClient.GetAgentsAsync(kind: badKind).ToListAsync();
+        List<ProjectsAgentRecord> test = [.. records.Where(x => string.Equals(x.Name, AGENT_NAME))];
+        Assert.That(test, Has.Count.EqualTo(0));
+        records = await agentsClient.GetAgentsAsync(kind: goodKind).ToListAsync();
+        test = [.. records.Where(x => string.Equals(x.Name, AGENT_NAME))];
+        Assert.That(test, Has.Count.EqualTo(1));
+        ValidateDefinition(test[0].Versions.Latest, useExternalAgent);
+        test = [.. records.Where(x => string.Equals(x.Name, AGENT_NAME2))];
+        Assert.That(test, Has.Count.EqualTo(1));
+        ValidateDefinition(test[0].Versions.Latest, useExternalAgent);
+
+        // List Versions
+        List<ProjectsAgentVersion> recordVersions = await agentsClient.GetAgentVersionsAsync(agentName: AGENT_NAME2).ToListAsync();
+        Assert.That(recordVersions, Has.Count.EqualTo(1));
+        ValidateDefinition(recordVersions[0], useExternalAgent);
+
+        recordVersions = await agentsClient.GetAgentVersionsAsync(agentName: AGENT_NAME).ToListAsync();
+        Assert.That(recordVersions, Has.Count.EqualTo(2));
+        ValidateDefinition(recordVersions[0], useExternalAgent);
+        // DeleteVersion
+        string expectedVersion = recordVersions[1].Version;
+        await agentsClient.DeleteAgentVersionAsync(agentName: AGENT_NAME, agentVersion: recordVersions[0].Version);
+        recordVersions = await agentsClient.GetAgentVersionsAsync(agentName: AGENT_NAME).ToListAsync();
+        Assert.That(recordVersions, Has.Count.EqualTo(1));
+        Assert.That(recordVersions[0].Version, Is.EqualTo(expectedVersion));
+        // Delete
+        await agentsClient.DeleteAgentAsync(agentName: AGENT_NAME2);
+        records = await agentsClient.GetAgentsAsync(kind: goodKind).Where(x => string.Equals(x.Name, AGENT_NAME2)).ToListAsync();
+        Assert.That(records, Has.Count.EqualTo(0));
+    }
+
+    private static void ValidateDefinition(ProjectsAgentVersion agent, bool useExternalAgent)
+    {
+        if (useExternalAgent)
+        {
+            Assert.That(agent.Definition, Is.InstanceOf<ExternalAgentDefinition>());
+        }
+        else
+        {
+            Assert.That(agent.Definition, Is.InstanceOf<DeclarativeAgentDefinition>());
+        }
     }
 
     [RecordedTest]
@@ -448,17 +490,21 @@ public class AgentsTests : AgentsTestBase
             VersionSelector = new([new FixedRatioVersionSelectionRule(agentVersion: agentVersion.Version, trafficPercentage: 74)]),
             Protocols = { AgentEndpointProtocol.Responses }
         };
-        AgentsSkill simpleSkill = await skillsClient.CreateSkillAsync(name: SKILL, description: "Calculates the sum of two numbers.", instructions: """
-            To calculate the sum  run
-            ```bash
-            echo $((<first> + <second>))
-            ```
-            ```powershell
-            (<first> + <second>)
-            ```
-            Replace <first> and <second> by the actual summation arguments.
-            """);
-        AgentCard card = new(version: "42", [new AgentCardSkill(id: simpleSkill.SkillId, name: SKILL)]);
+        SkillInlineContent content = new(
+            description: "Calculates the sum of two numbers.",
+            instructions: """
+                To calculate the sum  run
+                ```bash
+                echo $((<first> + <second>))
+                ```
+                ```powershell
+                (<first> + <second>)
+                ```
+                Replace <first> and <second> by the actual summation arguments.
+                """
+        );
+        SkillVersion simpleSkill = await skillsClient.CreateSkillVersionAsync(name: SKILL, inlineContent: content);
+        AgentCard card = new(version: "42", [new AgentCardSkill(id: simpleSkill.Id, name: SKILL)]);
         PatchAgentOptions patchOptions = new()
         {
             AgentEndpoint = config,
@@ -474,7 +520,7 @@ public class AgentsTests : AgentsTestBase
         Assert.That(((FixedRatioVersionSelectionRule)patchedRecord.AgentEndpoint.VersionSelector.VersionSelectionRules[0]).TrafficPercentage, Is.EqualTo(74));
         Assert.That(patchedRecord.AgentCard.Version, Is.EqualTo("42"));
         Assert.That(patchedRecord.AgentCard.Skills, Has.Count.EqualTo(1));
-        Assert.That(patchedRecord.AgentCard.Skills[0].Id, Is.EqualTo(simpleSkill.SkillId));
+        Assert.That(patchedRecord.AgentCard.Skills[0].Id, Is.EqualTo(simpleSkill.Id));
     }
 
     [RecordedTest]
@@ -659,9 +705,8 @@ public class AgentsTests : AgentsTestBase
         Assert.That(writeResponse.Path, Is.EqualTo($"storage/{file2}"));
         Assert.That(writeResponse.BytesWritten, Is.EqualTo(fileLength));
         // List
-        SessionDirectoryListResponse response = await filesClient.GetSessionFilesAsync(agentName: agentVersion.Name, sessionId: session.AgentSessionId, sessionStoragePath: "storage");
-        Assert.That(response.Entries, Has.Count.EqualTo(2));
-        List<string> lstEntries = [.. response.Entries.Select(x => x.Name)];
+        AsyncCollectionResult<SessionDirectoryEntry> response = filesClient.GetSessionFilesAsync(agentName: agentVersion.Name, agentSessionId: session.AgentSessionId, sessionStoragePath: "storage");
+        List<string> lstEntries = await response.Select(x => x.Name).ToListAsync();
         Assert.That(lstEntries, Does.Contain(file1));
         Assert.That(lstEntries, Does.Contain(file2));
         // Download
@@ -679,17 +724,65 @@ public class AgentsTests : AgentsTestBase
         Assert.That(data, Is.EqualTo("The test file\n"));
         // Delete
         await filesClient.DeleteSessionFileAsync(agentName: agentVersion.Name, sessionId: session.AgentSessionId, path: $"storage/{file2}");
-        response = await filesClient.GetSessionFilesAsync(agentName: agentVersion.Name, sessionId: session.AgentSessionId, sessionStoragePath: "storage");
-        lstEntries = [.. response.Entries.Select(x => x.Name)];
+        response = filesClient.GetSessionFilesAsync(agentName: agentVersion.Name, agentSessionId: session.AgentSessionId, sessionStoragePath: "storage");
+        lstEntries = await response.Select(x => x.Name).ToListAsync();
         Assert.That(lstEntries, Has.Count.EqualTo(1));
         Assert.That(lstEntries[0], Is.EqualTo(file1));
         await filesClient.DeleteSessionFileAsync(agentName: agentVersion.Name, sessionId: session.AgentSessionId, path: $"storage/{file1}");
-        response = await filesClient.GetSessionFilesAsync(agentName: agentVersion.Name, sessionId: session.AgentSessionId, sessionStoragePath: "storage");
-        Assert.That(response.Entries, Has.Count.EqualTo(0));
+        response = filesClient.GetSessionFilesAsync(agentName: agentVersion.Name, agentSessionId: session.AgentSessionId, sessionStoragePath: "storage");
+        Assert.That(await response.ToListAsync(), Has.Count.EqualTo(0));
         await agentsClient.DeleteSessionAsync(
             agentName: agentVersion.Name,
             sessionId: session.AgentSessionId
         );
+    }
+
+    [RecordedTest]
+    public async Task TestSessionFilePagination()
+    {
+        AgentAdministrationClient agentsClient = GetTestClient();
+        AgentSessionFiles filesClient = agentsClient.GetAgentSessionFiles();
+        ProjectsAgentVersion agentVersion = await CreateHostedAgent(agentsClient, "01");
+        ProjectAgentSession session = await agentsClient.CreateSessionAsync(
+            agentName: agentVersion.Name,
+            versionIndicator: new VersionRefIndicator(agentVersion.Version)
+        );
+        string fileLocalPath = GetTestFile("test.txt");
+        int fileLength = File.ReadAllBytes(fileLocalPath).Length;
+        // Make sure that chronological order is the reverse of session ID alphanumeric order.
+        for (int i = 0; i < PAGE_SIZE + 1; i++)
+        {
+            SessionFileWriteResponse writeResponse = await filesClient.UploadSessionFileAsync(
+                agentName: agentVersion.Name,
+                sessionId: session.AgentSessionId,
+                sessionStoragePath: $"storage/file{i}.json",
+                localPath: fileLocalPath
+            );
+            Assert.That(writeResponse.Path, Is.EqualTo($"storage/file{i}.json"));
+            Assert.That(writeResponse.BytesWritten, Is.EqualTo(fileLength));
+        }
+        List<SessionDirectoryEntry> records = await filesClient.GetSessionFilesAsync(agentName: agentVersion.Name, agentSessionId: session.AgentSessionId, sessionStoragePath: "storage", limit: PAGE_SIZE, order: AgentListOrder.Ascending).ToListAsync();
+        Assert.That(records.Count, Is.EqualTo(PAGE_SIZE + 1));
+        // Go forward.
+        //List<SessionDirectoryEntry> forward = await filesClient.GetSessionFilesAsync(agentName: agentVersion.Name, agentSessionId: session.AgentSessionId, sessionStoragePath: "storage", order: AgentListOrder.Ascending, after: records[0].Name, limit: PAGE_SIZE).ToListAsync();
+        //Assert.That(forward.Count, Is.EqualTo(records.Count - 1));
+        //Assert.That(forward[0].Name, Is.EqualTo(records[1].Name));
+        //Assert.That(forward[forward.Count - 1].Name, Is.EqualTo(records[records.Count - 1].Name));
+        //// Two limits:
+        //forward = await filesClient.GetSessionFilesAsync(agentName: agentVersion.Name, agentSessionId: session.AgentSessionId, sessionStoragePath: "storage", order: AgentListOrder.Ascending, after: records[0].Name, before: records[3].Name, limit: PAGE_SIZE).ToListAsync();
+        //Assert.That(forward.Count, Is.EqualTo(2));
+        //Assert.That(forward[0].Name, Is.EqualTo(records[1].Name));
+        //Assert.That(forward[1].Name, Is.EqualTo(records[2].Name));
+        //// Go backwards.
+        //List<SessionDirectoryEntry> backwards = await filesClient.GetSessionFilesAsync(agentName: agentVersion.Name, agentSessionId: session.AgentSessionId, sessionStoragePath: "storage", order: AgentListOrder.Descending, before: records[0].Name, limit: PAGE_SIZE).ToListAsync();
+        //Assert.That(backwards.Count, Is.EqualTo(records.Count - 1));
+        //Assert.That(backwards[0].Name, Is.EqualTo(records[records.Count - 1].Name));
+        //Assert.That(backwards[backwards.Count - 1].Name, Is.EqualTo(records[1].Name));
+        //// Two limits.
+        //backwards = await filesClient.GetSessionFilesAsync(agentName: agentVersion.Name, agentSessionId: session.AgentSessionId, sessionStoragePath: "storage", order: AgentListOrder.Descending, after: records[records.Count - 1].Name, before: records[records.Count - 4].Name, limit: PAGE_SIZE).ToListAsync();
+        //Assert.That(backwards.Count, Is.EqualTo(2));
+        //Assert.That(backwards[0].Name, Is.EqualTo(records[records.Count - 2].Name));
+        //Assert.That(backwards[1].Name, Is.EqualTo(records[records.Count - 3].Name));
     }
 
     [LiveOnly]
@@ -700,8 +793,8 @@ public class AgentsTests : AgentsTestBase
         AgentAdministrationClient agentsClient = GetTestClient();
         ProjectAgentSkills skillsClient = agentsClient.GetAgentSkills();
         string fileSkillName = $"{SKILL}_file";
-        AgentsSkill skillFromFile = await skillsClient.CreateSkillFromPackageAsync(GetTestFile("roll-dice"));
-        Assert.That(skillFromFile.Name, Is.EqualTo(fileSkillName));
+        AgentsSkill skillFromFile = await skillsClient.CreateSkillVersionFromFilesAsync("roll-dice", GetTestFile("roll-dice"));
+        Assert.That(skillFromFile.Name, Is.EqualTo("roll-dice"));
         Assert.That(skillFromFile.Description, Is.EqualTo("Roll dice using a random number generator."));
         string savePath = Path.Combine(Path.GetTempPath(), "saved_skill");
         try
@@ -709,7 +802,7 @@ public class AgentsTests : AgentsTestBase
             Directory.Delete(savePath, true);
         }
         catch { }
-        BinaryData data = await skillsClient.DownloadSkillAsync(skillFromFile.Name, savePath);
+        BinaryData data = await skillsClient.GetSkillContentAsync(skillFromFile.Name, savePath);
         Assert.That(savePath, Does.Exist);
         string skillMdPath = Path.Combine(savePath, "SKILL.md");
         Assert.That(skillMdPath, Does.Exist);
@@ -732,7 +825,7 @@ public class AgentsTests : AgentsTestBase
     {
         AgentAdministrationClient agentsClient = GetTestClient();
         ProjectAgentSkills skillsClient = agentsClient.GetAgentSkills();
-        string codeSkillName = $"{SKILL}_code", codeSkillName2 = $"{SKILL}_code2";
+        string codeSkillName = $"{SKILL}-code", codeSkillName2 = $"{SKILL}-code2";
         await foreach (AgentsSkill skill1 in skillsClient.GetSkillsAsync())
         {
             if (skill1.Name.StartsWith(SKILL))
@@ -740,19 +833,26 @@ public class AgentsTests : AgentsTestBase
                 await skillsClient.DeleteSkillAsync(skill1.Name);
             }
         }
-        AgentsSkill skillFromCode = await skillsClient.CreateSkillAsync(name: codeSkillName, description: "Calculates the sum of two numbers.", instructions: """
-            To calculate the sum run
-            ```bash
-            echo $((<first> + <second>))
-            ```
-            ```powershell
-            (<first> + <second>)
-            ```
-            Replace <first> and <second> by the actual summation arguments.
-            """);
+        SkillInlineContent content = new(
+            description: "Calculates the sum of two numbers.",
+            instructions: """
+                To calculate the sum run
+                ```bash
+                echo $((<first> + <second>))
+                ```
+                ```powershell
+                (<first> + <second>)
+                ```
+                Replace <first> and <second> by the actual summation arguments.
+                """
+        );
+        SkillVersion skillFromCode = await skillsClient.CreateSkillVersionAsync(name: codeSkillName, inlineContent: content);
+        string oldVersion = skillFromCode.Version;
         Assert.That(skillFromCode.Name, Is.EqualTo(codeSkillName));
         Assert.That(skillFromCode.Description, Is.EqualTo("Calculates the sum of two numbers."));
-        AgentsSkill skillFromCode2 = await skillsClient.CreateSkillAsync(name: codeSkillName2, description: "Divides one number by the other.", instructions: """
+        content = new(
+            description: "Divides one number by the other.",
+            instructions: """
             To calculate the division run
             ```bash
             echo $((<first> / <second>))
@@ -761,26 +861,37 @@ public class AgentsTests : AgentsTestBase
             (<first> + <second>)
             ```
             Replace <first> and <second> by the actual division arguments.
-            """);
+            """
+        );
+        SkillVersion skillFromCode2 = await skillsClient.CreateSkillVersionAsync(name: codeSkillName2, inlineContent: content);
         Assert.That(skillFromCode2.Name, Is.EqualTo(codeSkillName2));
         Assert.That(skillFromCode2.Description, Is.EqualTo("Divides one number by the other."));
         // Update
-        AgentsSkill skill = await skillsClient.UpdateSkillAsync(name: codeSkillName, description: "Calculates the product of two numbers.", instructions: """
-            To calculate the multiplication run
-            ```bash
-            echo $((<first> * <second>))
-            ```
-            ```powershell
-            (<first> * <second>)
-            ```
-            Replace <first> and <second> by the actual multiplication arguments.
-            """);
+        content = new(
+            description: "Calculates the product of two numbers.",
+            instructions: """
+                To calculate the multiplication run
+                ```bash
+                echo $((<first> * <second>))
+                ```
+                ```powershell
+                (<first> * <second>)
+                ```
+                Replace <first> and <second> by the actual multiplication arguments.
+                """
+        );
+        SkillVersion updatedVersion = await skillsClient.CreateSkillVersionAsync(name: codeSkillName, inlineContent: content);
+        AgentsSkill skill = await skillsClient.UpdateSkillAsync(name: codeSkillName, defaultVersion: updatedVersion.Version);
         Assert.That(skill.Name, Is.EqualTo(codeSkillName));
         Assert.That(skill.Description, Is.EqualTo("Calculates the product of two numbers."));
+        skill = await skillsClient.UpdateSkillAsync(name: codeSkillName, defaultVersion: oldVersion);
+        Assert.That(skill.Name, Is.EqualTo(codeSkillName));
+        Assert.That(skill.Description, Is.EqualTo("Calculates the sum of two numbers."));
+        await skillsClient.UpdateSkillAsync(name: codeSkillName, defaultVersion: updatedVersion.Version);
         // Get
-        skillFromCode = await skillsClient.GetSkillAsync(name: codeSkillName);
-        Assert.That(skillFromCode.Name, Is.EqualTo(codeSkillName));
-        Assert.That(skillFromCode.Description, Is.EqualTo("Calculates the product of two numbers."));
+        AgentsSkill retrievedSkill = await skillsClient.GetSkillAsync(name: codeSkillName);
+        Assert.That(retrievedSkill.Name, Is.EqualTo(codeSkillName));
+        Assert.That(retrievedSkill.Description, Is.EqualTo("Calculates the product of two numbers."));
         // List
         HashSet<string> skills = [.. await skillsClient.GetSkillsAsync().Select(x => x.Name).ToListAsync()];
         Assert.That(skills, Does.Contain(codeSkillName));
@@ -802,12 +913,9 @@ public class AgentsTests : AgentsTestBase
         AgentAdministrationClient agentsClient = GetTestClient();
         ProjectAgentSkills skillsClient = agentsClient.GetAgentSkills();
         int initialCount = (await skillsClient.GetSkillsAsync(limit: PAGE_SIZE, order: AgentListOrder.Ascending).ToListAsync()).Count;
-        for (int i = 0; i < PAGE_SIZE + 1; i++)
-        {
-            await skillsClient.CreateSkillAsync(
-                name: $"{SKILL}_{i}",
-                description: "Calculates the sum of two numbers.",
-                instructions: """
+        SkillInlineContent contrent = new(
+            description: "Calculates the sum of two numbers.",
+            instructions: """
                 To calculate the sum  run
                 ```bash
                 echo $((<first> + <second>))
@@ -817,30 +925,37 @@ public class AgentsTests : AgentsTestBase
                 ```
                 Replace <first> and <second> by the actual summation arguments.
                 """
+        );
+        for (int i = 0; i < PAGE_SIZE + 1; i++)
+        {
+            await skillsClient.CreateSkillVersionAsync(
+                name: $"{SKILL}-{i}",
+                inlineContent: contrent
             );
         }
         List<AgentsSkill> records = await skillsClient.GetSkillsAsync(limit: PAGE_SIZE, order: AgentListOrder.Ascending).ToListAsync();
         Assert.That(records.Count, Is.EqualTo(PAGE_SIZE + 1 + initialCount));
         // Go forward.
-        List<AgentsSkill> forward = await skillsClient.GetSkillsAsync(order: AgentListOrder.Ascending, after: records[0].SkillId, limit: PAGE_SIZE).ToListAsync();
+        List<AgentsSkill> forward = await skillsClient.GetSkillsAsync(order: AgentListOrder.Ascending, after: records[0].Id, limit: PAGE_SIZE).ToListAsync();
         Assert.That(forward.Count, Is.EqualTo(records.Count - 1));
-        Assert.That(forward[0].SkillId, Is.EqualTo(records[1].SkillId));
-        Assert.That(forward[forward.Count - 1].SkillId, Is.EqualTo(records[records.Count - 1].SkillId));
+        Assert.That(forward[0].Id, Is.EqualTo(records[1].Id));
+        Assert.That(forward[forward.Count - 1].Id, Is.EqualTo(records[records.Count - 1].Id));
         // Two limits:
-        forward = await skillsClient.GetSkillsAsync(order: AgentListOrder.Ascending, after: records[0].SkillId, before: records[3].SkillId, limit: PAGE_SIZE).ToListAsync();
+        forward = await skillsClient.GetSkillsAsync(order: AgentListOrder.Ascending, after: records[0].Id, before: records[3].Id, limit: PAGE_SIZE).ToListAsync();
         Assert.That(forward.Count, Is.EqualTo(2));
-        Assert.That(forward[0].SkillId, Is.EqualTo(records[1].SkillId));
-        Assert.That(forward[1].SkillId, Is.EqualTo(records[2].SkillId));
+        Assert.That(forward[0].Id, Is.EqualTo(records[1].Id));
+        Assert.That(forward[1].Id, Is.EqualTo(records[2].Id));
+        // Blocked by work item 5312533
         // Go backwards.
-        List<AgentsSkill> backwards = await skillsClient.GetSkillsAsync(order: AgentListOrder.Descending, before: records[0].SkillId, limit: PAGE_SIZE).ToListAsync();
-        Assert.That(backwards.Count, Is.EqualTo(records.Count - 1));
-        Assert.That(backwards[0].SkillId, Is.EqualTo(records[records.Count - 1].SkillId));
-        Assert.That(backwards[backwards.Count - 1].SkillId, Is.EqualTo(records[1].SkillId));
+        //List<AgentsSkill> backwards = await skillsClient.GetSkillsAsync(order: AgentListOrder.Descending, before: records[0].Id, limit: PAGE_SIZE).ToListAsync();
+        //Assert.That(backwards.Count, Is.EqualTo(records.Count - 1));
+        //Assert.That(backwards[0].Id, Is.EqualTo(records[records.Count - 1].Id));
+        //Assert.That(backwards[backwards.Count - 1].Id, Is.EqualTo(records[1].Id));
         // Two limits.
-        backwards = await skillsClient.GetSkillsAsync(order: AgentListOrder.Descending, after: records[records.Count - 1].SkillId, before: records[records.Count - 4].SkillId, limit: PAGE_SIZE).ToListAsync();
-        Assert.That(backwards.Count, Is.EqualTo(2));
-        Assert.That(backwards[0].SkillId, Is.EqualTo(records[records.Count - 2].SkillId));
-        Assert.That(backwards[1].SkillId, Is.EqualTo(records[records.Count - 3].SkillId));
+        //List<AgentsSkill> backwards = await skillsClient.GetSkillsAsync(order: AgentListOrder.Descending, after: records[records.Count - 1].Id, before: records[records.Count - 4].Id, limit: PAGE_SIZE).ToListAsync();
+        //Assert.That(backwards.Count, Is.EqualTo(2));
+        //Assert.That(backwards[0].Id, Is.EqualTo(records[records.Count - 2].Id));
+        //Assert.That(backwards[1].Id, Is.EqualTo(records[records.Count - 3].Id));
     }
 
     #region Helpers
@@ -878,7 +993,7 @@ public class AgentsTests : AgentsTestBase
                 { "AZURE_OPENAI_ENDPOINT", $"https://{accountId}.cognitiveservices.azure.com/" },
                 { "AZURE_OPENAI_CHAT_DEPLOYMENT_NAME", TestEnvironment.FOUNDRY_MODEL_NAME }
             },
-            Image = TestEnvironment.AGENT_DOCKER_IMAGE,
+            ContainerConfiguration = new(TestEnvironment.AGENT_DOCKER_IMAGE)
         };
         ProjectsAgentVersionCreationOptions creationOptions = new(agentDefinition);
         creationOptions.Metadata["enableVnextExperience"] = "true";