+++ sdk/dns/Azure.ResourceManager.Dns/src/Customization/Models/DnsDSRecordData.Serialization.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
 #nullable disable
@@ -45,7 +45,7 @@ namespace Azure.ResourceManager.Dns
             if (options.Format != "W" && Optional.IsDefined(SystemData))
             {
                 writer.WritePropertyName("systemData"u8);
-                JsonSerializer.Serialize(writer, SystemData);
+                writer.WriteObjectValue(SystemData, options);
             }
             writer.WritePropertyName("properties"u8);
             writer.WriteStartObject();
@@ -73,12 +73,12 @@ namespace Azure.ResourceManager.Dns
             if (Optional.IsDefined(TargetResource))
             {
                 writer.WritePropertyName("targetResource"u8);
-                JsonSerializer.Serialize(writer, TargetResource);
+                writer.WriteObjectValue(TargetResource, options);
             }
             if (Optional.IsDefined(TrafficManagementProfile))
             {
                 writer.WritePropertyName("trafficManagementProfile");
-                JsonSerializer.Serialize(writer, TrafficManagementProfile);
+                writer.WriteObjectValue(TrafficManagementProfile, options);
             }
             if (Optional.IsCollectionDefined(DnsDSRecords))
             {
@@ -101,7 +101,7 @@ namespace Azure.ResourceManager.Dns
 #else
                     using (JsonDocument document = JsonDocument.Parse(item.Value))
                     {
-                        JsonSerializer.Serialize(writer, document.RootElement);
+                        writer.WriteObjectValue(document.RootElement, options);
                     }
 #endif
                 }
@@ -171,7 +171,7 @@ namespace Azure.ResourceManager.Dns
                 }
                 if (property.NameEquals("systemData"u8))
                 {
-                    systemData = JsonSerializer.Deserialize<ResourceManager.Models.SystemData>(property.Value.GetRawText());
+                    systemData = ModelReaderWriter.Read<ResourceManager.Models.SystemData>(BinaryData.FromString(property.Value.GetRawText()), ModelSerializationExtensions.WireOptions, AzureResourceManagerDnsContext.Default);
                     continue;
                 }
                 if (property.NameEquals("properties"u8))
@@ -221,7 +221,7 @@ namespace Azure.ResourceManager.Dns
                             {
                                 continue;
                             }
-                            targetResource = JsonSerializer.Deserialize<WritableSubResource>(property0.Value.ToString());
+                            targetResource = ModelReaderWriter.Read<WritableSubResource>(BinaryData.FromString(property0.Value.GetRawText()), ModelSerializationExtensions.WireOptions, AzureResourceManagerDnsContext.Default);
                             continue;
                         }
                         if (property0.NameEquals("trafficManagementProfile"))
@@ -231,7 +231,7 @@ namespace Azure.ResourceManager.Dns
                                 property0.ThrowNonNullablePropertyIsNull();
                                 continue;
                             }
-                            trafficManagementProfile = JsonSerializer.Deserialize<WritableSubResource>(property0.Value.ToString());
+                            trafficManagementProfile = ModelReaderWriter.Read<WritableSubResource>(BinaryData.FromString(property0.Value.GetRawText()), ModelSerializationExtensions.WireOptions, AzureResourceManagerDnsContext.Default);
                             continue;
                         }
                         if (property0.NameEquals("DSRecords"u8))